backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-multipart==0.0.9
python-dotenv==1.0.1
langgraph==0.2.28
groq==0.31.0
reportlab==4.2.2
httpx==0.27.2
pytest==8.3.2
pytest-asyncio==0.23.8
